create user level middleware and policy

+++ routes/web.php
@@ -3,6 +3,7 @@ use App\Http\Controllers\Auth\LoginController;
 use App\Http\Controllers\Auth\RegisterController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\MovieController;
+use App\Http\Middleware\RoleAdmin;
 use Illuminate\Support\Facades\Auth;
 
 Route::get('/movies', [MovieController::class, 'index'])->name('movies.index');
@@ -13,10 +14,12 @@ Route::get('/movies/create', [MovieController::class, 'create'])->name('movies.c
 Route::post('/movies', [MovieController::class, 'store'])->name('movies.store')->middleware('auth');
 Route::get('/movies/{slug}', [MovieController::class, 'show'])->name('movies.show');
 
-Route::get('/movies/{slug}/edit', [MovieController::class, 'edit'])->name('movies.edit');
+Route::get('/movies/{slug}/edit', [MovieController::class, 'edit'])->name('movies.edit')->middleware('auth',RoleAdmin::class);
+
 Route::put('/movies/{slug}', [MovieController::class, 'update'])->name('movies.update');
 
-Route::delete('/movies/{slug}', [MovieController::class, 'destroy'])->name('movies.destroy');
+Route::delete('/movies/{slug}', [MovieController::class, 'destroy'])->name('movies.destroy')->middleware('auth');
+
 
 // routes/web.php
 Route::get('/test', function () {
